requirement.txt
autopep8==2.0.0
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.10.0
PyJWT==2.6.0
pymssql==2.2.7
six==1.16.0
SQLAlchemy==1.4.44
tomli==2.0.1
waitress==2.1.2
Werkzeug==2.2.2
